src/requirements.txt
Flask==3.0.3
mariadb
numpy<2
pandas
prophet
prometheus-client==0.20.0
psycopg2
py-healthcheck==1.10.1
pymssql
pysftp
python-dotenv==1.0.1
requests==2.31.0 # requests-pkcs12 breaks with requests 2.32.*
requests-pkcs12
SQLAlchemy
zipfile36
pymysql